+++ pages/theme_category_share.php
@@ -37,7 +37,7 @@ $linksuffixprev=implode('&',$linksuffixprev);
 $internal_share_only=checkperm("noex");
 
 # Process deletion of access keys
-if (getval("deleteaccess","")!="")
+if (getval("deleteaccess","")!="" && enforcePostRequest(false))
 	{
 	$ref=getvalescaped("ref","",true);	
 	delete_collection_access_key($ref,getvalescaped("deleteaccess",""));
@@ -49,6 +49,7 @@ include "../include/header.php";
 
 <div class="BasicsBox"> 
 <form method=post id="themeform" action="<?php echo $baseurl_short?>pages/theme_category_share.php<?php echo $linksuffix ?>" onsubmit="return CentralSpacePost(this,true)">
+<?php generateFormToken("themeform"); ?>
 <input type="hidden" name="generateurl" id="generateurl" value="">
 <p><a href='<?php echo $baseurl_short?>pages/themes.php<?php echo $linksuffixprev?>' onclick="return CentralSpaceLoad(this,true);"><?php echo LINK_CARET_BACK ?><?php echo $lang['backtothemes']?></a></p>
 <h1><?php echo $lang["share_theme_category"] . " - " . $themename?></h1>
